Emacs: fix auto-installation of missing packages
The problem with the code up to now was that the variable
package-archive-contents was non-nil, so package-refresh-contents never got
called.  However, simply replacing the conditional with an unconditional call to
package-refresh-contents would be sub-optimal, since updating the package list
from the various package repositories can take a while and unnecessarily delay
Emacs startup.

The solution I came up with is to call package-refresh-contents when a package
is encountered that is not yet installed.