PA: reset "autospawn" to its default value ("yes")
This is part of the solution to getting my old desktop PC to suspend correctly.

The problem is that the kernel module for my sound card (snd_ice1724) needs to
be unloaded first, which cannot be done as long as PA is accessing it.  Thus it
is necessary to terminate PA first.  However, with "autospawn = no", PA cannot
automatically restart, hence "autospawn" needs to be changed to its default
value of "yes" for sound to work after suspending.

I originally set "autospawn = no" to keep PA from trying to start while an
application that needed exclusive sound card access was running.  However, I do
not think that I have had that problem in quite a while.

Also, since PA now starts automatically, stop starting it from .xprofile.