PulseAudio: stop loading the mmkbd-evdev module
This module is only relevant when not using a desktop environment that takes
care of handling multimedia keys.  Furthermore, in the case of KDE, it even
appears to get in the way of KMix.

Leave the original lines commented out just in case I ever start using a
classic window manager again.