Vim: install VAM the usual way
Yeah, I *could* keep using the live ebuild I've been using until now, but I
kind of like the idea of not having to install an extra package just to get
add-on management in Vim.

This differs from the minimal setup in the VAM documentation in that I added
the "helptags" line, without which it is not possible to read the VAM
documentation from within Vim.