newsbeuter: drop the ESR blog
For every interesting tech article there is at least one completely bogus
political post, and I have realised that, for me, the interesting articles do
not make up for them.

It's not even that there isn't truth to them, but he seems to love rubbing
things in your face and going too far in a direction that pre-agrees with his
world view (usually in the form of snide remarks).  Frankly, there are better
way to get people to see other points of view, if that is even his goal.